requirements.txt
absl-py==0.7.1
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
cloudpickle==1.2.1
cycler==0.10.0
decorator==4.4.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
mock==3.0.5
networkx==2.2
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.2
six==1.12.0
sklearn==0.0
subprocess32==3.5.4
tb-nightly==1.14.0a20190614
tensorflow-probability==0.7.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019061401
tf-nightly-gpu==1.14.1.dev20190614
Werkzeug==0.15.4
wrapt==1.11.1
